Refactor rtcall-related and lift into ModuleContext.
This factors out from FunctionContext most of the rt-call-related
routines and lifts into ModuleContext. These have no FunctionContext
dependences and upcoming users need access without a FunctionContext.
