+++ language/tools/move-mv-llvm-compiler/src/stackless/translate.rs
@@ -39,7 +39,6 @@ use env_logger::fmt::Color;
 use log::{debug, Level};
 use move_core_types::{account_address, u256::U256, vm_status::StatusCode::ARITHMETIC_ERROR};
 use move_model::{ast as mast, model as mm, ty as mty};
-use move_native::shared::{MOVE_TYPE_DESC_SIZE, MOVE_UNTYPED_VEC_DESC_SIZE};
 use move_stackless_bytecode::{
     function_target::FunctionData, stackless_bytecode as sbc,
     stackless_bytecode_generator::StacklessBytecodeGenerator,
@@ -411,7 +410,7 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
                         builder.load_store(llty, src_llval, dst_llval);
                     }
                     mty::Type::Vector(elt_mty) => {
-                        self.emit_rtcall_with_retval(RtCall::VecCopy(
+                        self.module_cx.emit_rtcall_with_retval(RtCall::VecCopy(
                             dst_llval.as_any_value(),
                             src_llval.as_any_value(),
                             (**elt_mty).clone(),
@@ -615,7 +614,8 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
         // Generate the conditional branch and call to abort.
         builder.build_cond_br(cond_reg, then_bb, join_bb);
         builder.position_at_end(then_bb);
-        self.emit_rtcall_abort_raw(ARITHMETIC_ERROR as u64);
+        self.module_cx
+            .emit_rtcall_abort_raw(ARITHMETIC_ERROR as u64);
         builder.position_at_end(join_bb);
     }
 
@@ -859,7 +859,7 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
             pred == llvm::LLVMIntPredicate::LLVMIntEQ || pred == llvm::LLVMIntPredicate::LLVMIntNE
         );
 
-        let mut dst_reg = self.emit_rtcall_with_retval(RtCall::VecCmpEq(
+        let mut dst_reg = self.module_cx.emit_rtcall_with_retval(RtCall::VecCmpEq(
             src0_reg,
             src1_reg,
             vec_elt_cmp_mty.clone(),
@@ -911,8 +911,11 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
             pred == llvm::LLVMIntPredicate::LLVMIntEQ || pred == llvm::LLVMIntPredicate::LLVMIntNE
         );
 
-        let mut dst_reg =
-            self.emit_rtcall_with_retval(RtCall::StructCmpEq(src0_reg, src1_reg, cmp_mty.clone()));
+        let mut dst_reg = self.module_cx.emit_rtcall_with_retval(RtCall::StructCmpEq(
+            src0_reg,
+            src1_reg,
+            cmp_mty.clone(),
+        ));
 
         // The above produces equality, so invert if this is a not-equal comparison.
         if pred == llvm::LLVMIntPredicate::LLVMIntNE {
@@ -1451,7 +1454,7 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
         src: &[mast::TempIndex],
     ) {
         let types = mty::Type::instantiate_vec(types.to_vec(), self.type_params);
-        let typarams = self.get_rttydesc_ptrs(&types);
+        let typarams = self.module_cx.get_rttydesc_ptrs(&types);
 
         let dst_locals = dst.iter().map(|i| &self.locals[*i]).collect::<Vec<_>>();
         let src_locals = src.iter().map(|i| &self.locals[*i]).collect::<Vec<_>>();
@@ -1511,15 +1514,6 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
         } else {
             self.module_cx.llvm_builder.call(ll_fn, &src);
         }
-    }
-
-    fn get_rttydesc_ptrs(&self, types: &[mty::Type]) -> Vec<llvm::Constant> {
-        let mut ll_global_ptrs = vec![];
-        for type_ in types {
-            let ll_tydesc = self.module_cx.rtty_cx.define_llvm_tydesc(type_);
-            ll_global_ptrs.push(ll_tydesc.ptr());
-        }
-        ll_global_ptrs
     }
 
     fn translate_fun_call(
@@ -1738,7 +1732,9 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
         //   call move_rt_vec_copy(ptr @__move_rttydesc_{T}, %pv, @vec_data_descriptor)
         //   ...
 
-        let res_val = self.emit_rtcall_with_retval(RtCall::VecEmpty(elt_mty.clone()));
+        let res_val = self
+            .module_cx
+            .emit_rtcall_with_retval(RtCall::VecEmpty(elt_mty.clone()));
 
         // Be sure to emit allocas only in the entry block. They may otherwise be
         // interpreted as dynamic stack allocations by some parts of the LLVM code. These
@@ -1756,7 +1752,7 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
 
         builder.build_store(res_val, res_ptr);
 
-        self.emit_rtcall_with_retval(RtCall::VecCopy(
+        self.module_cx.emit_rtcall_with_retval(RtCall::VecCopy(
             res_ptr.as_any_value(),
             vec_descriptor.as_any_value(),
             elt_mty.clone(),
@@ -1767,7 +1763,7 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
     fn emit_rtcall(&self, rtcall: RtCall) {
         match &rtcall {
             RtCall::Abort(local_idx) => {
-                let llfn = self.get_runtime_function(&rtcall);
+                let llfn = self.module_cx.get_runtime_function(&rtcall);
                 let local_llval = self.locals[*local_idx].llval;
                 let local_llty = self.locals[*local_idx].llty;
                 self.module_cx.llvm_builder.load_call_store(
@@ -1778,8 +1774,8 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
                 self.module_cx.llvm_builder.build_unreachable();
             }
             RtCall::VecDestroy(local_idx, elt_mty) => {
-                let llfn = self.get_runtime_function(&rtcall);
-                let typarams = self.get_rttydesc_ptrs(&[elt_mty.clone()]);
+                let llfn = self.module_cx.get_runtime_function(&rtcall);
+                let typarams = self.module_cx.get_rttydesc_ptrs(&[elt_mty.clone()]);
                 let typarams = typarams.into_iter().map(|llval| llval.as_any_value());
                 // The C ABI passes the by-val-vector as a pointer.
                 let local = &self.locals[*local_idx];
@@ -1789,208 +1785,6 @@ impl<'mm, 'up> FunctionContext<'mm, 'up> {
             }
             _ => unreachable!(),
         }
-    }
-
-    // This version is used in contexts where TempIndexes are not used and/or where the caller
-    // expects a return value that it will decide how to use or store.
-    fn emit_rtcall_with_retval(&self, rtcall: RtCall) -> llvm::AnyValue {
-        match &rtcall {
-            RtCall::VecCopy(ll_dst_value, ll_src_value, elt_mty) => {
-                // Note, no retval from vec_copy.
-                let llfn = self.get_runtime_function(&rtcall);
-                let mut typarams: Vec<_> = self
-                    .get_rttydesc_ptrs(&[elt_mty.clone()])
-                    .iter()
-                    .map(|llval| llval.as_any_value())
-                    .collect();
-                typarams.push(*ll_dst_value);
-                typarams.push(*ll_src_value);
-                self.module_cx.llvm_builder.call(llfn, &typarams)
-            }
-            RtCall::VecCmpEq(ll_dst_value, ll_src_value, elt_mty) => {
-                let llfn = self.get_runtime_function(&rtcall);
-                let mut typarams: Vec<_> = self
-                    .get_rttydesc_ptrs(&[elt_mty.clone()])
-                    .iter()
-                    .map(|llval| llval.as_any_value())
-                    .collect();
-                typarams.push(*ll_dst_value);
-                typarams.push(*ll_src_value);
-                self.module_cx.llvm_builder.call(llfn, &typarams)
-            }
-            RtCall::VecEmpty(elt_mty) => {
-                let llfn = self.get_runtime_function(&rtcall);
-                let typarams: Vec<_> = self
-                    .get_rttydesc_ptrs(&[elt_mty.clone()])
-                    .iter()
-                    .map(|llval| llval.as_any_value())
-                    .collect();
-                self.module_cx.llvm_builder.call(llfn, &typarams)
-            }
-            RtCall::StructCmpEq(ll_src1_value, ll_src2_value, s_mty) => {
-                let llfn = self.get_runtime_function(&rtcall);
-                let mut typarams: Vec<_> = self
-                    .get_rttydesc_ptrs(&[s_mty.clone()])
-                    .iter()
-                    .map(|llval| llval.as_any_value())
-                    .collect();
-                typarams.push(*ll_src1_value);
-                typarams.push(*ll_src2_value);
-                self.module_cx.llvm_builder.call(llfn, &typarams)
-            }
-            _ => unreachable!(),
-        }
-    }
-
-    fn get_runtime_function(&self, rtcall: &RtCall) -> llvm::Function {
-        let name = match rtcall {
-            RtCall::Abort(..) => "abort",
-            RtCall::VecDestroy(..) => "vec_destroy",
-            RtCall::VecCopy(..) => "vec_copy",
-            RtCall::VecCmpEq(..) => "vec_cmp_eq",
-            RtCall::VecEmpty(..) => "vec_empty",
-            RtCall::StructCmpEq(..) => "struct_cmp_eq",
-        };
-        self.get_runtime_function_by_name(name)
-    }
-
-    fn get_runtime_function_by_name(&self, rtcall_name: &str) -> llvm::Function {
-        let fn_name = format!("move_rt_{rtcall_name}");
-        let llmod = &self.module_cx.llvm_module;
-        let llcx = &self.module_cx.llvm_cx;
-        let llfn = llmod.get_named_function(&fn_name);
-        if let Some(llfn) = llfn {
-            llfn
-        } else {
-            let (llty, attrs) = match rtcall_name {
-                "abort" => {
-                    let ret_ty = llcx.void_type();
-                    let param_tys = &[llcx.int_type(64)];
-                    let llty = llvm::FunctionType::new(ret_ty, param_tys);
-                    let attrs = vec![
-                        (llvm::LLVMAttributeFunctionIndex, "noreturn", None),
-                        (llvm::LLVMAttributeFunctionIndex, "cold", None),
-                    ];
-                    (llty, attrs)
-                }
-                "vec_destroy" => {
-                    // vec_destroy(type_ve: &MoveType, v: MoveUntypedVector)
-                    let ret_ty = llcx.void_type();
-                    let tydesc_ty = llcx.int_type(8).ptr_type();
-                    // The vector is passed by value, but the C ABI here passes structs by reference,
-                    // so it's another pointer.
-                    let vector_ty = llcx.int_type(8).ptr_type();
-                    let param_tys = &[tydesc_ty, vector_ty];
-                    let llty = llvm::FunctionType::new(ret_ty, param_tys);
-                    let attrs = self.mk_pattrs_for_move_type(1);
-                    (llty, attrs)
-                }
-                "vec_copy" => {
-                    // vec_copy(type_ve: &MoveType, dstv: &mut MoveUntypedVector, srcv: &MoveUntypedVector)
-                    let ret_ty = llcx.void_type();
-                    let tydesc_ty = llcx.int_type(8).ptr_type();
-                    // The vectors are passed by value, but the C ABI here passes structs by reference,
-                    // so it's another pointer.
-                    let vector_ty = llcx.int_type(8).ptr_type();
-                    let param_tys = &[tydesc_ty, vector_ty, vector_ty];
-                    let llty = llvm::FunctionType::new(ret_ty, param_tys);
-                    let mut attrs = self.mk_pattrs_for_move_type(1);
-                    attrs.extend(self.mk_pattrs_for_move_untyped_vec(2, true /* mut */));
-                    attrs.extend(self.mk_pattrs_for_move_untyped_vec(3, false /* !mut */));
-                    (llty, attrs)
-                }
-                "vec_cmp_eq" => {
-                    // vec_cmp_eq(type_ve: &MoveType, v1: &MoveUntypedVector, v2: &MoveUntypedVector) -> bool
-                    let ret_ty = llcx.int_type(1);
-                    let tydesc_ty = llcx.int_type(8).ptr_type();
-                    // The vectors are passed by value, but the C ABI here passes structs by reference,
-                    // so it's another pointer.
-                    let vector_ty = llcx.int_type(8).ptr_type();
-                    let param_tys = &[tydesc_ty, vector_ty, vector_ty];
-                    let llty = llvm::FunctionType::new(ret_ty, param_tys);
-                    let mut attrs = self.mk_pattrs_for_move_type(1);
-                    attrs.extend(self.mk_pattrs_for_move_untyped_vec(2, false /* !mut */));
-                    attrs.extend(self.mk_pattrs_for_move_untyped_vec(3, false /* !mut */));
-                    (llty, attrs)
-                }
-                "vec_empty" => {
-                    // vec_empty(type_ve: &MoveType) -> MoveUntypedVector
-                    let ret_ty = self.module_cx.get_llvm_type_for_move_native_vector();
-                    let tydesc_ty = llcx.int_type(8).ptr_type();
-                    let param_tys = &[tydesc_ty];
-                    let llty = llvm::FunctionType::new(ret_ty, param_tys);
-                    let attrs = self.mk_pattrs_for_move_type(1);
-                    (llty, attrs)
-                }
-                "struct_cmp_eq" => {
-                    // struct_cmp_eq(type_ve: &MoveType, s1: &AnyValue, s2: &AnyValue) -> bool;
-                    let ret_ty = llcx.int_type(1);
-                    let tydesc_ty = llcx.int_type(8).ptr_type();
-                    let anyval_ty = llcx.int_type(8).ptr_type();
-                    let param_tys = &[tydesc_ty, anyval_ty, anyval_ty];
-                    let llty = llvm::FunctionType::new(ret_ty, param_tys);
-                    let mut attrs = self.mk_pattrs_for_move_type(1);
-                    attrs.push((2, "readonly", None));
-                    attrs.push((2, "nonnull", None));
-                    attrs.push((3, "readonly", None));
-                    attrs.push((3, "nonnull", None));
-                    (llty, attrs)
-                }
-                n => panic!("unknown runtime function {n}"),
-            };
-
-            let ll_fn = llmod.add_function(&fn_name, llty);
-            llmod.add_attributes(ll_fn, &attrs);
-            ll_fn
-        }
-    }
-
-    fn mk_pattrs_for_move_type(
-        &self,
-        attr_idx: llvm::LLVMAttributeIndex,
-    ) -> Vec<(llvm::LLVMAttributeIndex, &'static str, Option<u64>)> {
-        assert!(
-            attr_idx != llvm::LLVMAttributeReturnIndex
-                && attr_idx != llvm::LLVMAttributeFunctionIndex
-        );
-        vec![
-            (attr_idx, "readonly", None),
-            (attr_idx, "nonnull", None),
-            (attr_idx, "dereferenceable", Some(MOVE_TYPE_DESC_SIZE)),
-        ]
-    }
-
-    fn mk_pattrs_for_move_untyped_vec(
-        &self,
-        attr_idx: llvm::LLVMAttributeIndex,
-        mutable: bool,
-    ) -> Vec<(llvm::LLVMAttributeIndex, &'static str, Option<u64>)> {
-        assert!(
-            attr_idx != llvm::LLVMAttributeReturnIndex
-                && attr_idx != llvm::LLVMAttributeFunctionIndex
-        );
-        let mut attrs = vec![
-            (attr_idx, "nonnull", None),
-            (
-                attr_idx,
-                "dereferenceable",
-                Some(MOVE_UNTYPED_VEC_DESC_SIZE),
-            ),
-        ];
-        if !mutable {
-            attrs.push((attr_idx, "readonly", None));
-        }
-        attrs
-    }
-
-    fn emit_rtcall_abort_raw(&self, val: u64) {
-        let thefn = self.get_runtime_function_by_name("abort");
-        let param_ty = self.module_cx.llvm_cx.int_type(64);
-        let const_llval = llvm::Constant::int(param_ty, U256::from(val));
-        self.module_cx
-            .llvm_builder
-            .build_call_imm(thefn, &[const_llval]);
-        self.module_cx.llvm_builder.build_unreachable();
     }
 }
